+++ R/zzz.R
@@ -7,7 +7,7 @@ startup <- function() {
   packageStartupMessage(
     "'tna' package version ", utils::packageVersion("tna"), "\n",
     "------------------------------------------------------\n",
-    "  Tikka, S., L\u00f3pez-Pernas, S. and Saqr, M. (2025). \n",
+    "  Tikka, S., L\u00f3pez-Pernas, S., and Saqr, M. (2025). \n",
     "  tna: An R Package for Transition Network Analysis.\n",
     "  Applied Psychological Measurement.\n",
     "  https://doi.org/10.1177/01466216251348840\n",